Extract test utility functions to test_util module
