+++ src/app_engine/analytics_test.py
@@ -1,50 +1,28 @@
 # Copyright 2014 Google Inc. All Rights Reserved.
 
-import unittest
-import webtest
-
-import analytics
-import apprtc
 import datetime
 import json
 import time
+import unittest
+import webtest
+
 from google.appengine.api import memcache
 from google.appengine.ext import testbed
 
-class ReplaceFunction(object):
-  """Makes it easier to replace a function in a class or module."""
-  def __init__(self, obj, function_name, new_function):
-    self.obj = obj
-    self.function_name = function_name
-    self.old_function = getattr(self.obj, self.function_name)
-    setattr(self.obj, self.function_name, new_function)
+import analytics
+from test_util import CapturingFunction
+from test_util import ReplaceFunction
 
-  def __del__(self):
-    setattr(self.obj, self.function_name, self.old_function)
-
-class CapturingFunction(object):
-  """Captures the last arguments called on a function."""
-  def __init__(self, retValue=None):
-    self.retValue = retValue
-    self.lastArgs = None
-    self.lastKwargs = None
-
-  def __call__(self, *args, **kwargs):
-    self.lastArgs = args
-    self.lastKwargs = kwargs
-
-    if callable(self.retValue):
-      return self.retValue()
-
-    return self.retValue
 
 class FakeBigQuery(object):
   """Handles long function calls to the Google API client."""
+
   def __init__(self):
     self.tabledata = CapturingFunction(self)
     self.insertAll = CapturingFunction(self)
     self.execute = CapturingFunction(
         {u'kind': u'bigquery#tableDataInsertAllResponse'})
+
 
 class AnalyticsTest(unittest.TestCase):
   """Test the Analytics class in the analytics module."""
